Refactor net as independent module under root
